fix: do not force tag names to lowercase
For certain use cases, we want to preserve the original casing of the
tags in the document. For purposes of matching/comparison, it is fine to
lowercase them internally, but we don't want to emit tags with the
original names modified. While we could "standardize" only tags which
_should_ be lowercased, this commit does not do that, and instead
removes the code which performs this transformation on the output.