attach spans to events
The main abstraction is the new function `Emitter::move_position`.
The `CallbackEmitter` uses this to track positions and spans.
For that, the tokenizer has to do some more work and emit more events,
so that the `Emitter` can track the position accurately.

An alternative was to add a `Reader` parameter to all emitter methods,
which can be used by `CallbackEmitter` to get spans from the reader.
We decided for this approach as it introduces less breaking changes and
clutter for the users which don't want to consume spans.