Fix compilation of fuzzer, make upgrade path for spans smoother, add CI (#126)
When spans were implemented, the fuzzer was not updated. Make it compile again. Upgrade a few dependencies, fix compilation, add a few helper impls in html5gum to make the upgrade path to 0.8.0+ less painful.

Fix CI so that it actually runs in the right directory too. Previously it'd cargo check from the main folder.

Also add a new mode in the fuzzer to catch bugs in spans.